LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-Document-write-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
abczxcvbnjkl

PASS document.write with html assigned via policy (successful URL transformation).
PASS document.write with multiple trusted arguments.
PASS document.writeln with html assigned via policy (successful URL transformation).
PASS document.writeln with multiple trusted arguments.
PASS `document.write(string)` throws
PASS `document.write(string, string)` throws
PASS `document.write(string, TrustedHTML)` throws
PASS `document.writeln(string)` throws
PASS `document.writeln(string, string)` throws
PASS `document.writeln(string, TrustedHTML)` throws
PASS `document.write(null)` throws
PASS `document.writeln(null)` throws
PASS `document.write(string)` observes default policy
PASS `document.write(string, string)` observes default policy
PASS `document.write(string, TrustedHTML)` observes default policy
PASS `document.writeln(string)` observes default policy
PASS `document.writeln(string, string)` observes default policy
PASS `document.writeln(string, TrustedHTML)` observes default policy
